Implemented viewPrereq into displayHelp and added JavaDocs to command classes.
