+++ src/main/java/seedu/classmate/commands/HelpCommand.java
@@ -2,16 +2,22 @@ package seedu.classmate.commands;
 
 import seedu.classmate.Display;
 import seedu.classmate.Major;
-import seedu.classmate.Parser;
 import seedu.classmate.SpecialisationOverview;
 
 /**
- * Displays the help message containing all commands to the user.
+ * Represents a Command that displays the list of available commands to the user.
  */
 public class HelpCommand extends Command {
 
+    /**
+     * Displays the help message containing all support commands.
+     *
+     * @param major The {@code Major} instance (unused).
+     * @param display The {@code Display} handler used to show messages.
+     * @param specialisationOverview The overview of specialisations (unused).
+     */
     @Override
     public void executeCommand(Major major, Display display, SpecialisationOverview specialisationOverview) {
-        Parser.displayHelp();
+        Display.displayHelp();
     }
 }
