+++ src/main/java/seedu/classmate/commands/PrereqCommand.java
@@ -6,14 +6,30 @@ import seedu.classmate.Major;
 import seedu.classmate.Module;
 import seedu.classmate.SpecialisationOverview;
 
+/**
+ * Represents a command that displays the prerequisite tree of a specified module.
+ */
 public class PrereqCommand extends Command{
 
     private final String moduleCode;
 
+    /**
+     * Constructs a {@code PrereqCommand} with the given module code.
+     *
+     * @param args The module code entered by the user.
+     */
     public PrereqCommand(String args) {
         this.moduleCode = args.trim();
     }
 
+    /**
+     * Displays the prerequisite tree of the specified module.
+     *
+     * @param major The {@code Major} instance containing module data.
+     * @param display The {@code Display} handler (unused).
+     * @param specialisationOverview The overview of specialisations (unused).
+     * @throws ClassMateException If no module code is provided.
+     */
     @Override
     public void executeCommand(Major major, Display display, SpecialisationOverview specialisationOverview) {
         if (moduleCode.isEmpty()) {
