+++ src/main/java/seedu/classmate/commands/QueryModuleAvailabilityCommand.java
@@ -6,11 +6,20 @@ import seedu.classmate.Major;
 import seedu.classmate.Module;
 import seedu.classmate.SpecialisationOverview;
 
+/**
+ * Represents a command that checks the availability of a module in a given semester.
+ */
 public class QueryModuleAvailabilityCommand extends Command {
 
     private final String moduleCode;
     private final String semester;
 
+    /**
+     * Constructs a {@code QueryModuleAvailabilityCommand} with the given arguments.
+     *
+     * @param args The user input containing the module code and semester.
+     * @throws ClassMateException If the input does not contain both module code and semester.
+     */
     public QueryModuleAvailabilityCommand(String args) {
 
         String[] parts = args.trim().split("\\s+");
@@ -23,6 +32,14 @@ public class QueryModuleAvailabilityCommand extends Command {
         this.semester = parts[1].toLowerCase();
     }
 
+    /**
+     * Checks and displays the availability of the specified module in a given semester.
+     *
+     * @param major The {@code Major} instance containing module data.
+     * @param display The {@code Display} handler (unused).
+     * @param specialisationOverview The overview of specialisations (unused).
+     * @throws ClassMateException If the module is not found.
+     */
     @Override
     public void executeCommand(Major major, Display display, SpecialisationOverview specialisationOverview) {
