+++ src/main/java/seedu/classmate/commands/SpecialisationInfoCommand.java
@@ -6,14 +6,30 @@ import seedu.classmate.Major;
 import seedu.classmate.Specialisation;
 import seedu.classmate.SpecialisationOverview;
 
+/**
+ * Represents a command that displays detailed information about a specific CEG specialisation.
+ */
 public class SpecialisationInfoCommand extends Command {
 
     private final String args;
 
+    /**
+     * Constructs a {@code SpecialisationInfoCommand} with the given arguments.
+     *
+     * @param args The user input representing the specialisation index.
+     */
     public SpecialisationInfoCommand(String args) {
         this.args = args.trim();
     }
 
+    /**
+     * Displays information about the specified specialisation.
+     *
+     * @param major The {@code Major} instance (unused).
+     * @param display The {@code Display} handler (unused).
+     * @param specialisationOverview The overview containing all specialisations.
+     * @throws ClassMateException If the input is empty or not a valid number (1 to 5).
+     */
     @Override
     public void executeCommand(Major major, Display display, SpecialisationOverview specialisationOverview) {
